requirements.txt
dill==0.3.3
ipython==7.21.0
isort==4.3.21
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
matplotlib==3.1.3
ncls==0.0.57
numpy==1.18.1
opencv-python==4.2.0.32
orjson==3.5.1
pandas==1.0.1
path==13.1.0
pathlib2==2.3.5
Pillow==7.0.0
py7zr==0.16.0
pydocstyle==4.0.1
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
pyyaml==5.4.1
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
sklearn==0.0
statsmodels==0.11.0
tensorboardX==2.1
torch==1.7.1
tqdm==4.42.1
